+++ src/app/components/Layout/Header/LoginComponents.tsx
@@ -8,11 +8,14 @@ export const LoginComponent = ({ me }: any) => {
   return (
     <div className="w-[100px]">
       {me?.user ? (
-        <img
-          alt="프로필"
-          src={me?.user?.image!}
-          className="rounded-[50%] w-[50px] h-[50px]"
-        />
+        <button>
+          <img
+            alt="프로필"
+            src={me?.user?.image!}
+            className="rounded-[50%] w-[50px] h-[50px]"
+            onClick={() => router.push("/my")}
+          />
+        </button>
       ) : (
         <Button
           variant="ivory"
