+++ src/app/login/_component/Button.tsx
@@ -1,0 +1,17 @@
+import { signIn } from "next-auth/react";
+
+const Button = () => {
+  const onClick = async () => {
+    await signIn("kakao", {
+      redirect: true,
+      callbackUrl: "/",
+    });
+  };
+  return (
+    <button className="border-2 p-2" onClick={onClick}>
+      카카오 로그인
+    </button>
+  );
+};
+
+export default Button;
