+++ src/app/login/page.tsx
@@ -1,17 +1,17 @@
-"use client";
-import { signIn } from "next-auth/react";
+import { auth } from "@/auth";
+import { redirect } from "next/navigation";
+import Button from "./_component/Button";
 
-export default function LoginPage() {
-  const onClick = async () => {
-    await signIn("kakao", {
-      redirect: true,
-      callbackUrl: "/",
-    });
-  };
+export default async function LoginPage() {
+  const session = await auth();
+
+  if (session?.user) {
+    redirect("/");
+  }
 
   return (
-    <button className="border-2 p-2" onClick={onClick}>
-      카카오 로그인
-    </button>
+    <div>
+      <Button />
+    </div>
   );
 }
